requirements.txt
fastapi
uvicorn
websocket-client
pydantic
python-dotenv
